perf: isolate can_parse, simple-absolute, and finish helpers
Move the SIMD percent-encode kernel, simple-absolute fast path,
relative resolution, and can_parse into their own TUs so the unity
ada.cpp inlining budget is not spent on those paths. Inline SSSE3
scanners, reuse the host delimiter, and drop npos cmovs.
