+++ include/ada/parser-inl.h
@@ -4,10 +4,10 @@
 #ifndef ADA_PARSER_INL_H
 #define ADA_PARSER_INL_H
 
+#include "ada/parser.h"
 #include "ada/expected.h"
 #include "ada/url_pattern.h"
 #include "ada/url_pattern_helpers.h"
-#include "ada/parser.h"
 
 #include <string>
 #include <string_view>
